feat: switch staking with NFTs link in popup nav
